+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.0.0 - 2023-05-XX
 * Added support for meridian window restriction
+* Added default exposure time to Exposure Templates
 * Added airmass to acquired image data detail display
 * Added option to park the mount while waiting for next target
 * Added option to throttle exposure planning when not grading